requirements.txt
flask
faiss-cpu
numpy
fastembed
waitress
sqlite3
requests